Point Gulp-Shell to my personal git.
The Gulp-Shell plugin is missing some features, and those features
are either still in review by the owner of this project, or not yet
in the npm version.

By moving Gulp-Shell to my personal directory (which contains the two
necessary updates for this project to work), it allows people to start
working on running those linters.

When the PR is merged (https://github.com/sun-zheng-an/gulp-shell/pull/20)
and when the npm package has been updated, we should update the endpoint.